backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
# Replace the old package with this new one:
google-genai